requirements.txt
blinker==1.7.0
bson==0.5.10
click==8.1.7
colorama==0.4.6
dnspython==2.5.0
Flask==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.1
six==1.16.0
Werkzeug==3.0.1
